+++ docs/changelog.rst
@@ -13,6 +13,7 @@ Changelog
 **Changed**
 
 - **Python**: Explicitly note that combining ``@schema.given`` with explicit examples from the spec is not supported. :issue:`1217`
+- Clarify error message when a state machine has no transitions. :issue:`1992`
 
 **Fixed**